app/backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-dotenv
python-multipart
SQLAlchemy
selenium
fastapi-mail
alembic
requests
beautifulsoup4
aiohttp
webdriver-manager>=3.8.6,<4.0.0
